# backend/Fms/openapi/index.yaml
openapi: 3.0.1
info:
  title: FMS
  version: v1
servers:
  - url: http://localhost:3333
paths:
...
  /api/organizations:
    put:
      tags:
        - Organization
      summary: Create new organization
      operationId: createOrganization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpsertRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpsertRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/OrganizationUpsertRequest'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Organization id
          content:
            application/json:
              schema:
                type: integer
                format: int32
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
    get:
      tags:
        - Organization
      summary: Get list of current user's organizations
      operationId: listUserOrganizations
      parameters:
        - name: Offset
          in: query
          schema:
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
  '/api/organizations/{id}':
    get:
      tags:
        - Organization
      summary: Get organization
      operationId: getOrganization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Organization data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
    post:
      tags:
        - Organization
      summary: Update organization
      operationId: updateOrganization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpsertRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpsertRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/OrganizationUpsertRequest'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
    delete:
      tags:
        - Organization
      summary: Delete organization
      operationId: deleteOrganization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
  '/api/organizations/{organizationId}/users/{userId}':
    put:
      tags:
        - Organization
      summary: Add user to organization
      operationId: organizationAddUser
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    delete:
      tags:
        - Organization
      summary: Remove user from organization
      operationId: organizationRemoveUser
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    get:
      tags:
        - Organization
      summary: Get user role in organization
      operationId: organizationGetUserRole
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationRole'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    post:
      tags:
        - Organization
      summary: Update user role in organization
      operationId: organizationUpdateUserRole
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRole'
          text/json:
            schema:
              $ref: '#/components/schemas/OrganizationRole'
          application/*+json:
            schema:
              $ref: '#/components/schemas/OrganizationRole'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
  '/api/organizations/{organizationId}/users':
    get:
      tags:
        - Organization
      summary: Get list of users in organization
      operationId: listOrganizationUsers
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: Query
          in: query
          schema:
            type: string
            example: Jane
          example: Jane
        - name: Offset
          in: query
          schema:
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationUserListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
...
components:
  schemas:
...
    OrganizationListResponse:
      required:
        - items
        - totalCount
      type: object
      properties:
        totalCount:
          type: integer
          format: int32
          example: 165
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationResponse'
      additionalProperties: false
    OrganizationResponse:
      required:
        - id
        - name
        - role
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 243
        name:
          minLength: 1
          type: string
          example: Cool Organization Inc.
        role:
          $ref: '#/components/schemas/OrganizationRole'
      additionalProperties: false
    OrganizationRole:
      enum:
        - OWNER
        - ADMIN
        - MEMBER
      type: string
...
    OrganizationUpsertRequest:
      required:
        - name
      type: object
      properties:
        name:
          maxLength: 255
          minLength: 3
          type: string
          example: New Organization Name
      additionalProperties: false
    OrganizationUserListResponse:
      required:
        - items
        - totalCount
      type: object
      properties:
        totalCount:
          type: integer
          format: int32
          example: 165
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationUserResponse'
      additionalProperties: false
    OrganizationUserResponse:
      required:
        - role
        - user
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        role:
          $ref: '#/components/schemas/OrganizationRole'
      additionalProperties: false
    PublicClientError:
      required:
        - description
      type: object
      properties:
        description:
          minLength: 1
          type: string
          example: Something bad happened
        validationErrors:
          type: array
          items:
            type: string
          nullable: true
          example:
            - Email can't contain less than 3 characters
            - Password can't be empty
      additionalProperties: false
    PublicError:
      required:
        - description
      type: object
      properties:
        description:
          minLength: 1
          type: string
          example: Something bad happened
      additionalProperties: false
...
    UserResponse:
      required:
        - firstName
        - id
        - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 25
        firstName:
          minLength: 1
          type: string
          example: John
        lastName:
          minLength: 1
          type: string
          example: Snow
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme.\n\nEnter 'Bearer' [space] and then your token in the text input below.\nExample: 'Bearer 12345abcdef'"
      name: Authorization
      in: header
security:
  - Bearer: [ ]